# build/src/swagger_pro.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "version": "1.0.0",
        "description": "API documentation using Swagger UI"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "summary": "User login using email or username",
                "tags": ["Auth"],
                "requestBody": {
                    "description": "Login with email or username and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "user_name_and_email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Login successful" },
                    "400": { "description": "Invalid credentials" }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "summary": "Register a new user",
                "tags": ["User"],
                "requestBody": {
                    "description": "Provide user details for registration",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User registered successfully" },
                    "400": { "description": "Invalid input data" }
                }
            }
        },
        "/api/v1/stream-data/ask-ai": {
            "post": {
                "summary": "Stream AI responses via SSE",
                "description": "Streams responses from an AI agent in real-time using Server-Sent Events.",
                "operationId": "streamGraphResponse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "messages": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "role": {
                                                    "type": "string",
                                                    "example": "user"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "From customer_trans table show me first 10 record with fields like gender , account balance , and customerid ,no other condition required"
                                                }
                                            },
                                            "required": ["role", "content"]
                                        }
                                    },
                                    "request_id": {
                                        "type": "string",
                                        "example": "4"
                                    },
                                    "thread_id": {
                                        "type": "string",
                                        "example": "4"
                                    }
                                },
                                "required": ["messages", "request_id", "thread_id"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stream started. The response is a stream of SSE messages.",
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "type": "string",
                                    "example": "data: { \"node\": \"starter_agent\", \"values\": { ... } }\\n\\n"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or missing fields.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean", "example": false },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean", "example": false },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}